Clear inputs when start to connect or on error
If there are many devices, after flashing the first one, a click on "Connect" does not clear the inputs. In some cases (not connected yet to another device, have problem with another device etc.) we think the connect is finished and believe the inputs are update, but it is not. Thus we should clear the inputs everytime we try to connect.