Allow display to set image data message size
Interpret the second and third bytes of the `01` response notification from the
display as a uint16le containing the display's requested image data message
size.

Prior to this commit, the expected value of the `01` response was `01 f4 00`,
and the image part size was hardcoded to 240 bytes.

Interpreted as a uint16le, `f4 00` = 244d, which is the message size for a 4
byte uint32le part number concatenated to 240 image part data bytes.

Note that I've made an assumption that both bytes should be interpreted as the
message size; I've only ever seen the first byte change (to 0x65), so it's
possible the size should be read as a uint8 followed by an unknown byte.

This fixes a bug that prevented uploading images from Chrome on an older
Macbook Air, which received `01 65 00` in response to the control command `01`,
while a newer Android device received the expected `01 f4 00` from the same
display.

Forcing the Mac which got `01 65 00` to send 240 byte image data chunks
transfers data, but fails to update the display. Sending 97d (= 0x65 - 4) byte
chunks from the Mac succeeds in updating the display.

Tested on a 2.9" BWR epaper display with settings:

 - 296x128 BWR
 - No dithering
 - No compression
 - Second color on
 - No mirroring